int64 for negativeloss and apply Mac CI, exclude .output.pb

+++ onnx/backend/test/case/node/acosh.py
@@ -17,12 +17,12 @@ class Acosh(Base):
             outputs=['y'],
         )
 
-        x = np.array([10, np.e, 1]).astype(np.float64)
-        y = np.arccosh(x).astype(np.float64)  # expected output [2.99322295,  1.65745449,  0.]
+        x = np.array([10, np.e, 1]).astype(np.float32)
+        y = np.arccosh(x)  # expected output [2.99322295,  1.65745449,  0.]
         expect(node, inputs=[x], outputs=[y],
                name='test_acosh_example')
 
-        x = np.random.uniform(1.0, 10.0, (3, 4, 5)).astype(np.float64)
-        y = np.arccosh(x).astype(np.float64)
+        x = np.random.uniform(1.0, 10.0, (3, 4, 5)).astype(np.float32)
+        y = np.arccosh(x)
         expect(node, inputs=[x], outputs=[y],
                name='test_acosh')
